% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf2svm}
\alias{opf2svm}
\title{Converts an OPF subGraph object to a LIBSVM file}
\usage{
opf2svm(data, outputFile)
}
\arguments{
\item{data}{The subGraph object}

\item{outputFile}{LIBSVM output file name}
}
\value{
`NULL`
}
\description{
Converts an OPF subGraph object to a LIBSVM file
}
\examples{
dataset <- opf_read_subGraph(system.file("extdata/boat.dat",package = "LibOPF"))
File <- file.path(tempdir(), "boat.svm")
opf2svm(dataset,File)
opf_check(File)

}
