% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf2txt}
\alias{opf2txt}
\title{Converts an OPF subGraph object to a text file}
\usage{
opf2txt(data, outputFile)
}
\arguments{
\item{data}{OPF subGraph object}

\item{outputFile}{Text output file name}
}
\value{
`NULL`
}
\description{
Converts an OPF subGraph object to a text file
}
\examples{
dataset <- opf_read_subGraph(system.file("extdata/boat.dat",package = "LibOPF"))
File <- file.path(tempdir(), "boat.txt")
opf2txt(dataset,File)

}
