% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_create_subGraph}
\alias{opf_create_subGraph}
\title{Creates an empty subGraph structure}
\usage{
opf_create_subGraph(nnodes)
}
\arguments{
\item{nnodes}{Number of nodes}
}
\value{
Returns an empty subGraph
}
\description{
Creates an empty subGraph structure
}
\examples{
EmptySubgraph <- opf_create_subGraph(10)

}
