% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_info}
\alias{opf_info}
\title{Gives information about the OPF file}
\usage{
opf_info(dataSet)
}
\arguments{
\item{dataSet}{The OPF file}
}
\value{
`NULL`
}
\description{
Gives information about the OPF file
}
\examples{
dat <- opf_read_subGraph(system.file("extdata/boat.dat",package = "LibOPF"))
opf_info(dat)

}
