% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_merge}
\alias{opf_merge}
\title{Merge subGraphs}
\usage{
opf_merge(dataSets)
}
\arguments{
\item{dataSets}{An array of subGraph objects}
}
\value{
Returns the merged subGraph object
}
\description{
Merge subGraphs
}
\examples{
dat <- opf_read_subGraph(system.file("extdata/boat.dat",package = "LibOPF"))
dat2 <- opf_read_subGraph(system.file("extdata/Z1LINE.dat",package = "LibOPF"))
dataSet <- opf_merge(c(dat,dat2))

}
