% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_normalize}
\alias{opf_normalize}
\title{Normalizes data for the OPF classifier}
\usage{
opf_normalize(dataSet)
}
\arguments{
\item{dataSet}{The subGraph object}
}
\value{
Returns the normalized subGraph
}
\description{
Normalizes data for the OPF classifier
}
\examples{
dataset <- opf_read_subGraph(system.file("extdata/boat.dat",package = "LibOPF"))
dat <- opf_normalize(dataset)

}
