% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_read_classification}
\alias{opf_read_classification}
\title{Reads a file which contains the nodes' predicted labels}
\usage{
opf_read_classification(file)
}
\arguments{
\item{file}{The file which contains the nodes' predicted labels}
}
\value{
Returns the predicted labels list
}
\description{
Reads a file which contains the nodes' predicted labels
}
\examples{
File <- system.file("extdata/classification.txt",package = "LibOPF")
classification <- opf_read_classification(File)

}
