% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_read_distances}
\alias{opf_read_distances}
\title{Reads a file which contains the precalculated distances}
\usage{
opf_read_distances(file)
}
\arguments{
\item{file}{The file which contains the distances matrix}
}
\value{
Returns the precalculated distances matrix
}
\description{
Reads a file which contains the precalculated distances
}
\examples{
distances <- opf_read_distances(system.file("extdata/distances.dat",package = "LibOPF"))

}
