% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_read_modelfile}
\alias{opf_read_modelfile}
\title{Reads a file which contains the learned model}
\usage{
opf_read_modelfile(file)
}
\arguments{
\item{file}{The file which contains the learned model}
}
\value{
Returns the learned model object
}
\description{
Reads a file which contains the learned model
}
\examples{
classifier <- opf_read_modelfile(system.file("extdata/classifier.opf",package = "LibOPF"))

}
