% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{opf_write_distances}
\alias{opf_write_distances}
\title{Writes into a file the precalculated distances computed by opf_distances function}
\usage{
opf_write_distances(distances, file)
}
\arguments{
\item{distances}{The matrix produced by the opf distances function}

\item{file}{The file name where you want to save the distances}
}
\value{
`NULL`
}
\description{
Writes into a file the precalculated distances computed by opf_distances function
}
\examples{
dat <- opf_read_subGraph(system.file("extdata/boat.dat",package = "LibOPF"))
dist <- opf_distance(dat,3,0)
opf_write_distances(dist, file.path(tempdir(), "distances.dat"))

}
