% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{svm2opf}
\alias{svm2opf}
\title{Converts a LIBSVM file to an OPF subGraph object}
\usage{
svm2opf(inputFile)
}
\arguments{
\item{inputFile}{LIBSVM input file}
}
\value{
Returns the OPF object
}
\description{
Converts a LIBSVM file to an OPF subGraph object
}
\examples{
dataset <- svm2opf(system.file("extdata/boat.svm",package = "LibOPF"))

}
