% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibOPF.R
\name{txt2opf}
\alias{txt2opf}
\title{Converts a text file to an OPF subGraph object}
\usage{
txt2opf(inputFile)
}
\arguments{
\item{inputFile}{Text input file}
}
\value{
Returns the OPF object
}
\description{
Converts a text file to an OPF subGraph object
}
\examples{
dataset <- txt2opf(system.file("extdata/boat.txt",package = "LibOPF"))

}
